+++ src/script/stores/uiStore.ts
@@ -16,6 +16,8 @@ import { isInputPatternValid } from './connectionStore';
 import { classifier } from './Stores';
 import Gesture from '../domain/stores/gesture/Gesture';
 import Axes from '../domain/Axes';
+import PersistantWritable from '../repository/PersistantWritable';
+import { DropdownOption } from '../../components/buttons/Buttons';
 
 let text: (key: string, vars?: object) => string;
 t.subscribe(t => (text = t));
@@ -121,13 +123,6 @@ export enum MicrobitInteractions {
   AB,
 }
 
-export const highlightedAxis = writable(Axes.X);
-
-const initialMicrobitInteraction: MicrobitInteractions = MicrobitInteractions.AB;
-
-export const microbitInteraction = writable<MicrobitInteractions>(
-  initialMicrobitInteraction,
-);
 
 export type ModelEntry = {
   id: string;
@@ -147,6 +142,29 @@ export const availableModels: ModelEntry[] = [
     label: 'KNN',
   },
 ];
+
+const defaultModel: ModelEntry | undefined = availableModels.find(
+  model => model.id === 'NN',
+);
+
+if (!defaultModel) {
+  throw new Error('Default model not found!');
+}
+// TODO: Should just be model id instead of dropdown option
+export const preferredModel = new PersistantWritable<DropdownOption>(
+  {
+    id: defaultModel.id,
+    label: defaultModel.label,
+  },
+  'prefferedModel',
+)
+export const highlightedAxis = writable(Axes.X);
+
+const initialMicrobitInteraction: MicrobitInteractions = MicrobitInteractions.AB;
+
+export const microbitInteraction = writable<MicrobitInteractions>(
+  initialMicrobitInteraction,
+);
 
 /**
  * Workaround for an unrecoverable reconnect failure due to a bug in chrome/chromium
